🥅 Fix error filters and split controllers
