+++ src/app.controller.spec.ts
@@ -1,5 +1,5 @@
 import { Test } from '@nestjs/testing';
-import { CommunityController } from './community/community.controller';
+import { CommunityController } from './community/controllers/community.controller';
 import { CommunityService } from './community/services/community.service';
 
 describe('AppController', () => {
