+++ src/auth/auth.controller.ts
@@ -1,12 +1,21 @@
-import { Body, Controller, Logger, Optional, Post } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Logger,
+  Optional,
+  Post,
+  UseFilters,
+} from '@nestjs/common';
 import { LoginDto } from './login.dto';
 import { AuthService } from './auth.service';
 import { Public } from '../common/decorators/public.decorator';
 import { ApiTags } from '@nestjs/swagger';
 import { LoginResponse } from './login-response.model';
+import { RainbowAuthHttpFilter } from '../common/filters/rainbow-auth-http.filter';
 
 @ApiTags('auth')
 @Controller('auth')
+@UseFilters(RainbowAuthHttpFilter)
 export class AuthController {
   constructor(
     private readonly authService: AuthService,
