+++ src/auth/auth.module.ts
@@ -2,16 +2,11 @@ import { Module } from '@nestjs/common';
 import { AuthController } from './auth.controller';
 import { HttpModule } from '@nestjs/axios';
 import { AuthService } from './auth.service';
-import { APP_FILTER } from '@nestjs/core';
-import { RainbowAuthHttpFilter } from '../common/filters/rainbow-auth-http.filter';
 
 @Module({
   imports: [HttpModule],
   controllers: [AuthController],
-  providers: [
-    AuthService,
-    { provide: APP_FILTER, useClass: RainbowAuthHttpFilter },
-  ],
+  providers: [AuthService],
   exports: [AuthService],
 })
 export class AuthModule {}
