+++ src/community/community.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { CommunityController } from './community.controller';
+import { CommunityController } from './controllers/community.controller';
 import { CommunityService } from './services/community.service';
 import { TripService } from './services/trip.service';
 import { SequelizeModule } from '@nestjs/sequelize';
@@ -8,8 +8,7 @@ import { CommunityEntity } from '../common/entities/community.entity';
 import { UserTripEntity } from '../common/entities/user-trip.entity';
 import { UserEntity } from '../common/entities/user.entity';
 import { UserCommunityEntity } from '../common/entities/user-community.entity';
-import { APP_FILTER } from '@nestjs/core';
-import { RainbowHttpFilter } from '../common/filters/rainbow-http.filter';
+import { TripController } from './controllers/trip.controller';
 
 @Module({
   imports: [
@@ -21,11 +20,7 @@ import { RainbowHttpFilter } from '../common/filters/rainbow-http.filter';
       UserCommunityEntity,
     ]),
   ],
-  controllers: [CommunityController],
-  providers: [
-    CommunityService,
-    TripService,
-    { provide: APP_FILTER, useClass: RainbowHttpFilter },
-  ],
+  controllers: [CommunityController, TripController],
+  providers: [CommunityService, TripService],
 })
 export class CommunityModule {}
