+++ src/main.ts
@@ -6,6 +6,7 @@ import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import { AppConfig } from './app.config';
 import { INestApplication, LogLevel, ValidationPipe } from '@nestjs/common';
 import { SequelizeFilter } from './common/filters/sequelize.filter';
+import { RainbowHttpFilter } from './common/filters/rainbow-http.filter';
 
 async function bootstrap(): Promise<void> {
   const app = await NestFactory.create(AppModule);
@@ -19,7 +20,7 @@ async function bootstrap(): Promise<void> {
   logger.setLogLevels(logLevels);
   app.useGlobalPipes(new ValidationPipe());
   app.useLogger(logger);
-  app.useGlobalFilters(new SequelizeFilter());
+  app.useGlobalFilters(new RainbowHttpFilter(), new SequelizeFilter());
   const enableSwagger = configService.get('enableSwagger', { infer: true });
   if (enableSwagger) {
     enableSwaggerEndpoint(app);
